+++ Engine/src/Kargono/Network/Session.cpp
@@ -11,7 +11,7 @@ namespace Kargono::Network
 
 		// Notify Clients that session initialization has started
 		Kargono::Network::Message newMessage;
-		newMessage.Header.ID = CustomMsgTypes::CurrentSessionInit;
+		newMessage.Header.ID = static_cast<uint32_t>(CustomMsgTypes::CurrentSessionInit);
 		for (auto& [clientID, connection] : m_ConnectedClients)
 		{
 			connection->Send(newMessage);
@@ -23,7 +23,7 @@ namespace Kargono::Network
 		KG_INFO("[SERVER]: Starting to determine connection latencies");
 
 		// Record Current Time for each connection and send a ping to each client
-		newMessage.Header.ID = CustomMsgTypes::InitSyncPing;
+		newMessage.Header.ID = static_cast<uint32_t>(CustomMsgTypes::InitSyncPing);
 		for (auto& [clientID, connection] : m_ConnectedClients)
 		{
 			m_InitCache.LatencyCache.insert({ clientID, {} });
@@ -67,7 +67,7 @@ namespace Kargono::Network
 
 		// If Cache is not filled yet, send another sync ping
 		Kargono::Network::Message newMessage;
-		newMessage.Header.ID = CustomMsgTypes::InitSyncPing;
+		newMessage.Header.ID = static_cast<uint32_t>(CustomMsgTypes::InitSyncPing);
 		m_InitCache.RecentTimePoints.insert_or_assign(clientID, std::chrono::high_resolution_clock::now());
 		m_ConnectedClients.at(clientID)->Send(newMessage);
 	}
@@ -139,7 +139,7 @@ namespace Kargono::Network
 		// Send start message to each client
 		float waitTime{ 0 };
 		Kargono::Network::Message newMessage;
-		newMessage.Header.ID = CustomMsgTypes::StartSession;
+		newMessage.Header.ID = static_cast<uint32_t>(CustomMsgTypes::StartSession);
 		for (auto& [clientID, connection] : m_ConnectedClients)
 		{
 			// Calculate wait time
@@ -184,7 +184,7 @@ namespace Kargono::Network
 			// Send start message to each client
 			float waitTime{ 0 };
 			Kargono::Network::Message newMessage;
-			newMessage.Header.ID = CustomMsgTypes::SessionReadyCheckConfirm;
+			newMessage.Header.ID = static_cast<uint32_t>(CustomMsgTypes::SessionReadyCheckConfirm);
 			for (auto& [clientID, connection] : m_ConnectedClients)
 			{
 				// Calculate wait time
